changed variables to doubles

+++ main.cpp
@@ -46,17 +46,17 @@ int main( int argc, char * argv[] )
 	cout << loan_amount << " " << yearly_interest_rate << " " << monthly_payment << endl;
 
 	//get the monthly 
-	int Rate = yearly_interest_rate / 12;
+	double Rate = yearly_interest_rate / 12;
 
 	// interest_payment
-	int interest_payment = (Rate*loan_amount)/100;
+	double interest_payment = (Rate*loan_amount)/100;
 
 	// calculate the principle
-	int principle = monthly_payment - interest_payment;
+	double principle = monthly_payment - interest_payment;
 
 	int month = 0;
 
-	int total_interest = 0;
+	double total_interest = 0;
 
 	cout << endl;
 	// AMORTIZATION TABLE
